cdk-bedrock-agent/requirements.txt
aws-cdk-lib
constructs
aws-cdk.aws-lambda-python-alpha
cdklabs-generative-ai-cdk-constructs
# aws-cdk.aws-s3-deployment removed; use aws_cdk.aws_s3_deployment from aws-cdk-lib
